refactored cyclops keyboard project (#4734)
* refactored cyclops keyboard project

Moved the files into a subfolder to allow consistency in preparation for additional commits to come in the future.

made some small adjustments to the configuration and keymap

* Update keyboards/westfoxtrot/cyclops/keymaps/default/keymap.c


* Update keyboards/westfoxtrot/cyclops/readme.md


* Update keyboards/westfoxtrot/cyclops/readme.md


* Update keyboards/westfoxtrot/cyclops/readme.md
