+++ keyboards/westfoxtrot/cyclops/cyclops.c
@@ -1,4 +1,4 @@
-/* Copyright 2018 'mechmerlin'
+/* Copyright 2018 westfoxtrot
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
