+++ keyboards/westfoxtrot/cyclops/readme.md
@@ -1,13 +1,15 @@
-# Cyclops
+# cyclops
 
-66% Alps Keyboard in ISO layout. 
+![cyclops](https://github.com/westfoxtrot/Cyclops)
 
-Keyboard Maintainer: [MechMerlin](https://github.com/mechmerlin)  
-Hardware Supported: 66% Alps Keyboard  
-Hardware Availability: A very limited number were made. 
+66% Alps Keyboard in ISO layout.
+
+ Keyboard Maintainer: [westfoxtrot](https://github.com/westfoxtrot)  
+ Hardware Supported: Cyclops (AEKISO66) PCB  
+ Hardware Availability: Files released here (https://github.com/westfoxtrot/Cyclops)  
 
 Make example for this keyboard (after setting up your build environment):
 
-    make cyclops:default
+    make westfoxtrot/cyclops:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
